requirements.txt
Django==4.1.7
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==22.1
django-simple-history==3.2.0
djangorestframework==3.14.0
psycopg2==2.9.5
pytest-django==4.5.2

